+++ CHANGELOG.rst
@@ -4,9 +4,10 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(slices-difference): difference function added
 * 🐛 fix: git scripts now function properly
 * ✅ test(slice-compact): made test cases and benchmarks for compact function
-* 🎉 feat(slices-compact): Compact function added
+* 🎉 feat(slices-compact): compact function added
 * ✅ test(slice-chunk): made test cases and benchmarks more accurate
 * 🎉 feat(slice-chunk): chunk function added, updated, and optimized